More changes to Jenkinsfile

+++ Jenkinsfile
@@ -103,8 +103,10 @@ pipeline {
                     sh '''
                     ssh -i /home/jenkinsadmin/.ssh/id_rsa -o StrictHostKeyChecking=no ubuntu@18.206.125.73 <<EOF
                     sudo docker pull atulvsharma/helloworld:1.0
-                    sudo docker stop helloworld-container || true
-                    sudo docker rm helloworld-container || true
+                    if sudo docker ps -a --format '{{.Names}}' | grep -q 'helloworld-container'; then
+                       sudo docker stop helloworld-container || true
+                       sudo docker rm helloworld-container || true
+		    fi
                     sudo docker run -d -p 80:80 --name hellowrold-container atulvsharma/helloworld:latest
                     EOF
                     '''
